+++ pages/upload/index.jsx
@@ -1,0 +1,5 @@
+function Upload() {
+  return <div>Upload</div>
+}
+
+export default Upload
