+++ src/components/footer.jsx
@@ -1,0 +1,7 @@
+const Footer = () => (
+  <footer>
+    Made with ❤️in India🇮🇳
+  </footer>
+)
+
+export default Footer
